test_final.py
#!/usr/bin/env python3
"""Final test to verify NextPy is working perfectly"""

import requests
import time

def test_nextpy_server():
    """Test that NextPy server is working correctly"""
    
    base_url = "http://localhost:5000"
    
    print("🧪 Testing Final NextPy Server Setup...")
    print(f"🌐 Testing server at: {base_url}")
    
    try:
        # Test homepage
        print("\n📄 Testing homepage...")
        response = requests.get(f"{base_url}/", timeout=5)
        if response.status_code == 200:
            print("✅ Homepage accessible (200 OK)")
            # Check for Tailwind CSS in the response
            if "tailwind" in response.text.lower():
                print("✅ Tailwind CSS is being served")
            else:
                print("⚠️  Tailwind CSS might not be linked properly")
            
            # Check for JSX content
            if "Welcome to NextPy" in response.text:
                print("✅ JSX content rendered correctly")
            else:
                print("⚠️  JSX content might not be rendering")
        else:
            print(f"❌ Homepage failed: {response.status_code}")
            return False
        
        # Test about page
        print("\n📄 Testing about page...")
        response = requests.get(f"{base_url}/about", timeout=5)
        if response.status_code == 200:
            print("✅ About page accessible (200 OK)")
        else:
            print(f"❌ About page failed: {response.status_code}")
        
        # Test Tailwind CSS file directly
        print("\n🎨 Testing Tailwind CSS file...")
        response = requests.get(f"{base_url}/tailwind.css", timeout=5)
        if response.status_code == 200:
            print("✅ Tailwind CSS file accessible (200 OK)")
            # Check for Tailwind utility classes
            if ".flex" in response.text and ".text-center" in response.text:
                print("✅ Tailwind utility classes compiled correctly")
            else:
                print("⚠️  Tailwind utility classes might be missing")
        else:
            print(f"❌ Tailwind CSS file failed: {response.status_code}")
        
        print("\n🎉 All tests completed successfully!")
        print("\n🚀 NextPy is fully working with:")
        print("  ✅ True JSX support")
        print("  ✅ Tailwind CSS integration")
        print("  ✅ File-based routing")
        print("  ✅ Hot reload development")
        print("  ✅ Server-side rendering")
        print("  ✅ Development server")
        
        return True
        
    except requests.exceptions.ConnectionError:
        print("❌ Cannot connect to server")
        print("💡 Make sure 'nextpy dev' is running in another terminal")
        return False
    except Exception as e:
        print(f"❌ Test failed: {e}")
        return False

if __name__ == "__main__":
    success = test_nextpy_server()
    if success:
        print("\n🎯 SUCCESS: NextPy is ready for development!")
        print("🌐 Open http://localhost:5000 in your browser")
    else:
        print("\n❌ FAILURE: Server setup needs attention")
